flutter: Rebase onto the Flutter 2.8 project comes with lots of changes
Rebase the Flutter boilerplate code to the latest by creating a new
Flutter project with 2.8.0 version. But I have to update the running
Flutter version to 2.8.1 because on 2.8.0 I cannot build properly on my
Apple Silicon Macbook Pro.

Plus, I updated all the codes to be null-safety, which is modern and has
advantages to avoid null exception errors.

Last, here are about the packages:
- Relaced settings_ui to flutter_settings_ui because settings_ui seems
  like deprecated
- Update Firebase packages

Signed-off-by: Deokgyu Yang <secugyu@gmail.com>